requirements.txt
asgiref==3.2.10
attrs==19.3.0
awsebcli==3.18.1
bcrypt==3.1.7
blessed==1.17.5
boto==2.49.0
boto3==1.14.5
botocore==1.15.49
cached-property==1.5.1
cement==2.8.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
Django==3.0.5
django-cors-headers==3.2.1
django-filter==2.2.0
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docker==4.2.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
drf-yasg==1.17.1
future==0.16.0
idna==2.10
importlib-metadata==1.6.0
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.9.5
jsonschema==3.2.0
MarkupSafe==1.1.1
mysqlclient==1.4.6
packaging==20.4
paramiko==2.7.1
pathspec==0.5.9
Pillow==7.1.2
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.3.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
semantic-version==2.5.0
six==1.15.0
sqlparse==0.3.1
termcolor==1.1.0
texttable==1.6.2
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.1.9
websocket-client==0.57.0
zipp==3.1.0
